+++ GameScript.Language/Lexer/TokenType.cs
@@ -22,6 +22,7 @@
 		Range,
 		OpenBracket,
 		CloseBracket,
-		Dot
+		Dot,
+		Ellipsis
 	}
 }